requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aiosqlite==0.20.0
alembic==1.13.3
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
apify_client==1.8.1
apify_shared==1.1.2
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
auth0-python==4.7.2
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.35.76
botocore==1.35.76
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cohere==5.11.1
coloredlogs==15.0.1
Crawl4AI==0.3.731
cryptography==43.0.3
dataclasses-json==0.6.7
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docstring-parser==0.15
docx2txt==0.8
duckduckgo_search==6.3.5
durationpy==0.9
email_validator==2.2.0
emoji==2.14.0
eval_type_backport==0.2.0
executing==2.1.0
fake-http-header==0.3.5
fastapi==0.115.4
fastapi-cli==0.0.6
fastavro==1.9.7
filelock==3.16.1
filetype==1.2.0
firecrawl-py==1.5.0
flatbuffers==24.3.25
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.21.0
google-auth==2.35.0
google-cloud-aiplatform==1.70.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
google_search_results==2.4.2
googleapis-common-protos==1.65.0
googlesearch-python==1.2.5
gptcache==0.1.44
greenlet==3.0.3
grpc-google-iam-v1==0.13.1
grpcio==1.67.1
grpcio-status==1.62.3
grpcio-tools==1.62.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html2text==2024.2.26
html5lib==1.1
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.1
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
iniconfig==2.0.0
ipython==8.29.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.7.0
jmespath==1.0.1
joblib==1.4.2
json_repair==0.30.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpickle==3.3.0
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lancedb==0.14.0
langchain==0.3.9
langchain-community==0.3.9
langchain-core==0.3.21
langchain-openai==0.2.11
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.2
langdetect==1.0.9
langsmith==0.1.137
litellm==1.52.9
llama-cloud==0.1.4
llama-index==0.9.27
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core==0.11.22
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.4.0
llama-index-legacy==0.9.48.post4
llama-index-llms-openai==0.2.16
llama-index-multi-modal-llms-openai==0.2.3
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-llama-parse==0.3.0
llama-parse==0.5.13
lxml==5.3.0
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mem0ai==0.1.24
mmh3==5.0.1
mockito==1.5.3
monotonic==1.6
more-itertools==10.5.0
motor==3.6.0
mpmath==1.3.0
multidict==6.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
narwhals==1.10.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
nodeenv==1.9.1
oauthlib==3.2.2
olefile==0.47
openai==1.76.2
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
orjson==3.10.11
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.3
parameterized==0.9.0
parso==0.8.4
peewee==3.17.8
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
playwright==1.47.0
pluggy==1.5.0
portalocker==2.10.1
posthog==3.7.0
primp==0.7.0
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==4.25.5
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-extra-types==2.10.0
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydeck==0.9.1
pyee==12.0.0
Pygments==2.18.0
PyJWT==2.9.0
pylance==0.18.2
pymongo==4.9.2
PyMuPDF==1.24.13
pymupdf4llm==0.0.10
pypdf==4.3.1
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyright==1.1.386
pysbd==0.3.4
PySocks==1.7.1
pytest==8.3.3
pytest-mockito==0.0.4
python-dateutil==2.8.2
python-dotenv==1.0.1
python-http-client==3.3.7
python-iso639==2024.10.22
python-magic==0.4.27
python-multipart==0.0.17
python-oxmsg==0.0.1
pytube==15.0.0
pytz==2024.2
pyvis==0.3.2
PyYAML==6.0.2
qdrant-client==1.12.1
RapidFuzz==3.10.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
resend==2.4.0
retry==0.9.2
rich==13.9.4
rich-toolkit==0.11.3
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.4
schema==0.7.7
selenium==4.25.0
sendgrid==6.11.0
setuptools==75.3.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
starkbank-ecdsa==2.2.0
starlette==0.41.2
streamlit==1.39.0
striprtf==0.0.26
sympy==1.13.3
tabulate==0.9.0
tavily-python==0.5.0
tenacity==8.5.0
tf-playwright-stealth==1.1.0
tiktoken==0.8.0
tokenizers==0.20.1
toml==0.10.2
tomli_w==1.1.0
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
typer==0.15.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
uv==0.4.27
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
wikipedia==1.4.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.17.1
yfinance==0.2.49
youtube-transcript-api==0.6.2
zipp==3.20.2
